@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "parcel src/index.html  --open --https ",
+    "start": "parcel serve src/index.html  --open --https",
     "serve": "serve dist",
     "dev": "yarn start",
     "prebuild": "rimraf dist",
@@ -16,32 +16,31 @@
   },
   "dependencies": {
     "@auth0/auth0-react": "^1.4.0",
-    "@chakra-ui/react": "^1.8.8",
-    "@emotion/css": "^11.1.3",
+    "@chakra-ui/react": "^2.0.0",
+    "@emotion/css": "^11.9.0",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
+    "@fontsource/inter": "^4.5.10",
     "@reduxjs/toolkit": "^1.5.1",
     "@sentry/react": "^6.9.0",
     "@sentry/tracing": "^6.9.0",
     "@stripe/react-stripe-js": "^1.2.2",
     "@stripe/stripe-js": "^1.16.0",
-    "@types/react-redux": "^7.1.18",
-    "@types/redux": "^3.6.0",
     "axios": "^0.21.1",
     "canvas-confetti": "^1.3.2",
-    "framer-motion": "^6.2.10",
-    "hamburger-react": "^2.4.0",
-    "react": "^17.0.0",
+    "framer-motion": "^6.3.3",
+    "hamburger-react": "^2.5.0",
+    "react": "^18.1.0",
     "react-awesome-button": "^6.5.1",
     "react-circular-progressbar": "^2.0.4",
-    "react-cookie-consent": "^6.2.3",
+    "react-cookie-consent": "^7.4.1",
     "react-countdown-circle-timer": "^2.5.1",
     "react-device-detect": "^1.15.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-icons": "^4.3.1",
     "react-paypal-js": "^0.0.1-security",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0",
     "react-router-sitemap": "^1.2.0",
     "react-tooltip": "^4.2.17",
     "react-transition-group": "^4.4.1",
@@ -51,25 +50,16 @@
     "stripe": "^8.164.0"
   },
   "devDependencies": {
-    "@parcel/packager-raw-url": "2.0.0-beta.3.1",
-    "@parcel/transformer-image": "2.0.0-beta.3.1",
-    "@parcel/transformer-sass": "2.0.0-beta.3.1",
-    "@storybook/addon-actions": "^6.3.5",
-    "@storybook/addon-essentials": "^6.3.5",
-    "@storybook/addon-links": "^6.3.5",
-    "@storybook/preset-scss": "^1.0.3",
-    "@storybook/preset-typescript": "^3.0.0",
-    "@storybook/react": "^6.3.5",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.8",
-    "autoprefixer": "^10.3.1",
-    "parcel": "2.0.0-beta.3.1",
-    "parcel-config-pwa-manifest": "^0.0.9",
-    "parcel-reporter-static-files-copy": "^1.2.2",
-    "postcss": "^8.3.6",
+    "@parcel/packager-raw-url": "^2.5.0",
+    "@parcel/transformer-image": "^2.5.0",
+    "@parcel/transformer-sass": "^2.5.0",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@types/redux": "^3.6.0",
+    "parcel": "^2.5.0",
+    "parcel-config-pwa-manifest": "^0.1.2",
+    "parcel-reporter-static-files-copy": "^1.3.4",
     "react-snap": "^1.23.0",
-    "sass": "^1.37.5",
     "typescript": "^4.3.5",
     "typescript-plugin-css-modules": "^3.4.0"
   },
