@@ -1,6 +1,6 @@
 import Sitemap from 'react-router-sitemap';
 import * as Path from 'path';
-import routes from '../src/app/routes';
+import routes from '../src/layout/routes';
 
 const buildPath = Path.resolve(__dirname, '../dist/sitemap.xml');
 const host = 'http://fuku.tv';
