@@ -1,4 +1,5 @@
-import { createAsyncThunk, EnhancedStore } from '@reduxjs/toolkit';
+import type { EnhancedStore } from '@reduxjs/toolkit';
+import { createAsyncThunk } from '@reduxjs/toolkit';
 import * as fukuService from 'src/services/fukuService';
 
 interface Stats {
